Assignment 3

+++ payment/events/payment_completed.go
@@ -5,7 +5,7 @@ import "time"
 type PaymentCompletedEvent struct {
 	EventID       string    `json:"event_id"`
 	OrderID       string    `json:"order_id"`
-	Amount        float64   `json:"amount"`
+	Amount        int64     `json:"amount"`
 	CustomerEmail string    `json:"customer_email"`
 	Status        string    `json:"status"`
 	OccurredAt    time.Time `json:"occurred_at"`
